Ensure null-termination of filenames copied by strncpy
Prior use of strncpy passed the buffer length for n; when the string to copy is
length n or longer, this results in the buffer not containing a null byte. This
is problematic as the buffers are assumed to contain a null byte elsewhere.

Reported here: https://forums.freertos.org/t/freertos-fat-compiler-warnings/14843