Correct default definition of a config macro
The check for defining the default value of
ffconfigUNICODE_UTF16_SUPPORT was #if( ffconfigUNICODE_UTF16_SUPPORT == 0 )
which resulted in macro redefinition warnings if the application defined
the macro to 0. This commit changes the check to
 #if !defined( ffconfigUNICODE_UTF16_SUPPORT ) to ensure that default
value if only defined when the application does not define the macro and
hence, no warning is generated even when the application defines
ffconfigUNICODE_UTF16_SUPPORT to 0.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>